degrade ntpdate error to warning
It's not worth raising an error just because the server
connection was not sufficiently quick once.
We see this on our machines every now and then.

If this is a permanent error condition it could escalate at some point
but that just adds additional complexity here.
A warning should alert you enough if it doesn't go away and you observe
unusual behavior.